Added config option to toggle html-min usage
Currently it's only possible to minify the output if the production flag is set. Thus it is not possible to add the minification to the development system. As in HTML to have spaces between elements can result in different display results it should be possible to enable the minification manually to also respect this while developing and visually testing the CSS styles outside of the production environment.

The cleanest and most-backward compatible way I could think of is to just add the proposed 'enabled' configuration, which defaults to the current behavior of 'config.production', to toggle filters manually on and off.